+++ src/stores/PostStore.js
@@ -35,7 +35,6 @@ class PostStore extends EventEmitter {
       let {settingResult, body} = this.parsePostSetting(data.posts[0].body);
       this.removeTopTag(data.posts[0].tags);
       this.post = {...data.posts[0], body, image, ...settingResult};
-
       this.emitChange();
     }
   }
@@ -70,7 +69,12 @@ class PostStore extends EventEmitter {
     let imageSrc = body.match(/<img [^>]*\/>/g);
 
     if (imageSrc) {
-      imageSrc = imageSrc[0].match(/http[^\"\'\s]*/)[0];
+      let imageSrcList = imageSrc[0].match(/http[^\"\'\s]*/);
+      if (imageSrcList) {
+        imageSrc = imageSrcList[0];
+      } else {
+        imageSrc = 'https://fbcdn-sphotos-f-a.akamaihd.net/hphotos-ak-xpa1/t31.0-8/11154796_779010435550566_7018350735656129504_o.jpg';
+      }
     } else {
       imageSrc = 'https://fbcdn-sphotos-f-a.akamaihd.net/hphotos-ak-xpa1/t31.0-8/11154796_779010435550566_7018350735656129504_o.jpg';
     }